Bug fix  [import data file] with --start-clean reseting the metaDB tables of live migration in cleanImportState() only in live migration

+++ yb-voyager/cmd/importData.go
@@ -544,14 +544,16 @@ func cleanImportState(state *ImportDataState, tasks []*ImportFileTask) {
 		}
 	}
 
-	// clearing state from metaDB based on importerRole
-	err := metaDB.ResetQueueSegmentMeta(importerRole)
-	if err != nil {
-		utils.ErrExit("failed to reset queue segment meta: %s", err)
-	}
-	err = metaDB.DeleteJsonObject(identityColumnsMetaDBKey)
-	if err != nil {
-		utils.ErrExit("failed to reset identity columns meta: %s", err)
+	if changeStreamingIsEnabled(importType) { 
+		// clearing state from metaDB based on importerRole
+		err := metaDB.ResetQueueSegmentMeta(importerRole)
+		if err != nil {
+			utils.ErrExit("failed to reset queue segment meta: %s", err)
+		}
+		err = metaDB.DeleteJsonObject(identityColumnsMetaDBKey)
+		if err != nil {
+			utils.ErrExit("failed to reset identity columns meta: %s", err)
+		}
 	}
 }
